diet plan project/AI DIET/backend/requirements.txt
flask
flask-cors
python-dotenv
langchain-groq
langchain-core
flask-sqlalchemy
flask-bcrypt
flask-jwt-extended
